+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changed
 
 - Splitted the `download` command into `new`, `open`, and `download`. ([#3](https://github.com/qryxip/cargo-compete/pull/3))
+- Changed the format of `workspace-metadata.toml`.
+    - Now `cargo-compete.test-suite`s are [Liquid](https://shopify.github.io/liquid/) templates.  ([#4](https://github.com/qryxip/cargo-compete/pull/4))
 
 ### Fixed